requirements.txt
alembic==1.14.1
Authlib==1.3.2
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
cryptography==44.0.1
distlib==0.3.9
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.16.1
firebase-admin==6.6.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-Dance==7.1.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
pytest==8.3.4
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sendgrid==6.11.0
six==1.17.0
SQLAlchemy==2.0.38
starkbank-ecdsa==2.2.0
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
URLObject==2.4.3
virtualenv==20.29.2
Werkzeug==2.3.8
zipp==3.20.2
